moved taylor_expansion.py
