+++ dictos/calculus/finite_difference.py
@@ -14,7 +14,7 @@ from dictos.utils import (
 )
 from dictos.linalg.linalg import dot_product, div
 from dictos.poly.lagrangian_polynomial import lagrangian_poly, derivative
-from dictos.taylor_expansion import taylor_series, derivative_symbol
+from dictos.series.taylor_expansion import taylor_series, derivative_symbol
 from dictos.error.finite_difference import UnsupportedOrderOfDerivativeError
 from dictos.core.expr import Expr
