+++ dictos/poly/interpolation.py
@@ -14,7 +14,7 @@ from dictos.utils import (
 )
 from dictos.linalg.linalg import dot_product
 from dictos.poly.lagrangian_polynomial import lagrangian_poly
-from dictos.taylor_expansion import taylor_series
+from dictos.series.taylor_expansion import taylor_series
 from dictos.error.stencil import ContainsZeroError
 from dictos.core.expr import Expr
